--- creation/static/colors.css
@@ -1,234 +1,0 @@
-.timeline-container .timeline .container div, #rules-container .rules-card, #index-container .challenge {
-  box-shadow: 2px 8px 4px 4px rgba(0, 0, 0, 0.2);
-}
-
-body {
-  background-image:url('../../static/images/challstatement_background.png');
-}
-
-.justified-par {
-  text-align: justify;
-  text-justify: inter-word;
-}
-
-#navbar {
-  background-color: whitesmoke;
-}
-#navbar * {
-  background-color: transparent;
-  color: #121212;
-}
-#navbar .dropdown-content a {
-  color: black;
-  padding: 12px 16px;
-  text-decoration: none;
-  display: block;
-  background-color: white;
-}
-
-#index-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  align-items: center;
-  height: 100vh;
-}
-#index-container .challenge {
-  padding: 20px;
-  border-radius: 25px;
-  background-color: #121212;
-  color: whitesmoke;
-  height: 300px;
-}
-#index-container .challenge h3 {
-  height: 60px;
-}
-#index-container .challenge hr {
-  border: 1px solid whitesmoke;
-}
-#index-container .challenge:hover {
-  margin-top: -10px;
-  margin-bottom: 10px;
-}
-
-.bottom {
-  padding-top: 300px;
-  margin: 0;
-  background-color: transparent;
-}
-
-#regis-container {
-  width: 80%;
-  margin: 40px 10%;
-}
-@media (min-width: 768px) {
-  #regis-container {
-    width: 50%;
-    margin: 40px 25%;
-  }
-}
-#regis-container li {
-  color: black;
-}
-#regis-container small {
-  color: black;
-  filter: contrast(1000%);
-}
-
-#rules-container {
-  background-color: transparent;
-  padding-bottom: 50px;
-}
-#rules-container .col-lg-6 {
-  padding: 25px;
-}
-#rules-container .rules-card {
-  margin: 100px 10px;
-  background-color: #00BBF9;
-  padding: 20px;
-  border-radius: 10px;
-}
-#rules-container hr {
-  border: 1px solid black;
-}
-
-#faq-container {
-  background-color: transparent;
-  display: flex;
-  justify-content: center;
-  color: whitesmoke;
-}
-#faq-container .faq-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  align-items: baseline;
-}
-#faq-container .faq-container div {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  align-items: baseline;
-  margin-bottom: 30px;
-}
-#faq-container .faq-box {
-  margin-top: 100px;
-  width: 60%;
-  min-width: 300px;
-}
-#faq-container hr {
-  border: 1px solid whitesmoke;
-}
-
-.timeline-container {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  padding: 200px 0 50px 0;
-}
-.timeline-container #timeline-line {
-  position: absolute;
-  height: 100%;
-  background-color: whitesmoke;
-  width: 0.5em;
-  top: 0;
-}
-.timeline-container .timeline {
-  position: relative;
-  max-width: 1200px;
-  margin: 0 auto;
-}
-.timeline-container .timeline .timeline::after {
-  content: "";
-  position: absolute;
-  width: 6px;
-  background-color: white;
-  top: 0;
-  bottom: 0;
-  left: 50%;
-  margin-left: -3px;
-}
-.timeline-container .timeline .container {
-  padding: 10px 40px;
-  position: relative;
-  background-color: inherit;
-  width: 50%;
-}
-.timeline-container .timeline .container::after {
-  content: "";
-  position: absolute;
-  width: 25px;
-  height: 25px;
-  right: -17px;
-  background-color: white;
-  border: 4px solid #FF9F55;
-  top: 15px;
-  border-radius: 50%;
-  z-index: 1;
-}
-.timeline-container .timeline .left {
-  left: 0%;
-}
-.timeline-container .timeline .right {
-  left: 50%;
-}
-.timeline-container .timeline .left::before {
-  content: " ";
-  height: 0;
-  position: absolute;
-  top: 22px;
-  width: 0;
-  z-index: 1;
-  right: 30px;
-  border: medium solid white;
-  border-width: 10px 0 10px 10px;
-  border-color: transparent transparent transparent white;
-}
-.timeline-container .timeline .right::before {
-  content: " ";
-  height: 0;
-  position: absolute;
-  top: 22px;
-  width: 0;
-  z-index: 1;
-  left: 30px;
-  border: medium solid white;
-  border-width: 10px 10px 10px 0;
-  border-color: transparent white transparent transparent;
-}
-.timeline-container .timeline .right::after {
-  left: -16px;
-}
-.timeline-container .timeline .content {
-  padding: 20px 30px;
-  background-color: white;
-  position: relative;
-  border-radius: 6px;
-}
-@media screen and (max-width: 600px) {
-  .timeline-container .timeline .timeline::after {
-    left: 31px;
-  }
-  .timeline-container .timeline .container {
-    width: 100%;
-    padding-left: 70px;
-    padding-right: 25px;
-  }
-  .timeline-container .timeline .container::before {
-    left: 60px;
-    border: medium solid white;
-    border-width: 10px 10px 10px 0;
-    border-color: transparent white transparent transparent;
-  }
-  .timeline-container .timeline .left::after, .timeline-container .timeline .right::after {
-    left: 15px;
-  }
-  .timeline-container .timeline .right {
-    left: 0%;
-  }
-}
-.timeline-container .timeline p {
-  text-align: left;
-}
-
-/*# sourceMappingURL=colors.css.map */
