--- creation/templates/creation/base.html
@@ -1,82 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-    <!-- Required meta tags -->
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-        <!-- Bootstrap CSS -->
-        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
-        <link rel="stylesheet" href="../../../static/style.css">
-        <link rel="stylesheet" href="../../static/colors.css">
-        <title>CREATION 2021</title>
-    </head>
-    <body>
-        <div class="container">
-            <div class="row" id="navbar" style="display: flex;">
-                <div class="col-md-auto menu">
-                    <button class="dropbtn"><a href="/" class="dropbtn">Home</a></button>
-                </div>
-                <div class="col-md-auto menu">
-                    <div class="dropdown">
-                        <button class="dropbtn">Competition Details</button>
-                        <div class="dropdown-content">
-                            <a href={% url 'creation:challenge' %}>Challenge Statements</a>
-                            <a href={% url 'creation:registration' %}>Registration</a>
-                            <a href={% url 'creation:submission' %}>Submission Portal</a>
-                        </div>
-                      </div>
-                </div>
-                <div class="col-md-auto menu">
-                    <button class="dropbtn"><a href={% url 'creation:rules' %} class="dropbtn">Rules, Terms & Conditions</a></button>
-                </div>
-                <div class="col-md-auto menu">
-                    <button class="dropbtn"><a href={% url 'creation:faq' %} class="dropbtn">FAQs</a></button>
-                </div>
-                <div class="col-md-auto menu">
-                    <button class="dropbtn"><a href={% url 'creation:contact' %} class="dropbtn">Contact Us</a></button>
-                </div>
-                <div class="col-md-auto menu" style="margin-left: auto;">
-                    <div class="content-section">
-                        <div class="media">
-                            {% if user.username %}
-                            <img class="rounded-circle account-img" style="width: 40px; height: 40px; margin-top: 8px;"src="https://pbs.twimg.com/profile_images/1506592811/Untitled.png">
-                            <div class="media-body">
-                                <h3 class="dropbtn" className="username">{{user.username}}</h3>
-                            </div>
-                            <div class="media-body"> 
-                                <button class="dropbtn">
-                                    <a href={% url 'profile' %} class="dropbtn">Profile</a>
-                                </button>
-                            </div>
-                            <div class="media-body"> 
-                                <button class="dropbtn">
-                                    <a href={% url 'logout' %}>Logout</a>
-                                </button>
-                            </div>
-                            {% else %}
-                            <div class="media-body"> 
-                                <button class="dropbtn">
-                                    <a href={% url 'login' %} class="dropbtn">Log in</a>
-                                </button>
-                            </div>
-                            {% endif %}
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div>
-            {% if messages %}
-                {% for message in messages %}
-                    <div class="alert alert-{{ message.tags }}">
-                        {{ message }}
-                    </div>
-                {% endfor %}
-            {% endif %}
-        </div>
-        {% block content %}{% endblock %}
-        <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ho+j7jyWK8fNQe+A12Hb8AhRq26LrZ/JpcUGGOn+Y7RsweNrtN/tE3MoK7ZeZDyx" crossorigin="anonymous"></script>
-    </body>
-</html>
