+++ creation2021/urls.py
@@ -15,17 +15,14 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
-from users import views as users_views
 from django.contrib.auth import views as auth_views
 
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('', include(('creation.urls', 'creation'), namespace='creation')), 
     path('', include(('users.urls', 'users'), namespace='users')),
-    path('profile/', users_views.profile, name='profile'),
-    path('login/', auth_views.LoginView.as_view(template_name='users/login.html'), name='login'),
-    path('logout/', auth_views.LogoutView.as_view(template_name='users/logout.html'), name='logout'),
+    path('admin/', admin.site.urls), 
+    path('login/', auth_views.LoginView.as_view(template_name='users/backend/login.html'), name='login'),
+    path('logout/', auth_views.LogoutView.as_view(template_name='users/backend/logout.html'), name='logout'),
 
     #path('register/', users_views.register)
 ]
