+++ users/templates/users/backend/login.html
@@ -14,7 +14,7 @@
     </form>
     <div class="border-top pt-3">
         <small class="text-muted">
-            Haven't Made an Account? <a class="ml-2" href={% url 'creation:registration' %}>Sign Up</a>
+            Haven't Made an Account? <a class="ml-2" href={% url 'users:registration' %}>Sign Up</a>
         </small>
     </div>
 </div>
