+++ users/templates/users/frontend/challenge.html
@@ -1,4 +1,4 @@
-{% extends "creation/base.html" %}
+{% extends "users/base.html" %}
 {% block content %}
 <h1>Challenge</h1>
 {% endblock content %}
