+++ users/templates/users/frontend/contact.html
@@ -1,4 +1,4 @@
-{% extends "creation/base.html" %}
+{% extends "users/base.html" %}
 {% block content %}
 <div id='contactContainer'>
