+++ users/templates/users/frontend/faq.html
@@ -1,4 +1,4 @@
-{% extends "creation/base.html" %}
+{% extends "users/base.html" %}
 {% block content %}
 <div class="faqer" id="faq-container">
     <div class="faq-box">
