+++ users/templates/users/frontend/index.html
@@ -1,4 +1,4 @@
-{% extends "creation/base.html" %}
+{% extends "users/base.html" %}
 {% block content %}
 <div class="row bottom justify-content-md-center">
     <div class="col-md-3 challenge">
