+++ users/templates/users/frontend/rules.html
@@ -1,4 +1,4 @@
-{% extends "creation/base.html" %}
+{% extends "users/base.html" %}
 {% block content %}
 <div class="row justify-content-lg-center bottom" id="rules-container">
     <div class="col-lg-6">
