--- users/templates/users/user.html
@@ -1,14 +1,0 @@
-{% extends "users/base.html" %}
-{% block content %}
-<div class="container">
-    <h1>Hello, {{ user.first_name }}</h1>
-    <li>Currently logged in as: {{ user.username }}</li>
-    <li><a href={% url 'logout' %}>Logout</a></li> 
-</div>
-{% endblock content %}
-
-
-
-
-
-
